Added auto-height and refresh on resize
The height option can now have the value "auto" (which also is
default). With this, the column size is calculated so that the columns
fit in the browser (not vertical scrolling is necessary).

Furthermore, when using auto height, the columnization can now refresh
itself when the browser window is resized.  For this, the script now
duplicates the content and saves the original content in a hidden
element. On refresh, the content from the hidden element is pulled and
the old columns are replaced.

Furthermore, the following new options were introduced:
  - maxHeight: Determines the maximum height of columns when using auto
height, default is -1 (no limit)
  - columWidth: The width of a single column in pixels, default is 200
  - columnGap: The gap between two columns in pixels, default is 20
  - splitAccuracy: The accuracy when a text node is split, lower values
mean higher accuracy